+++ client/components/index.js
@@ -7,4 +7,4 @@ export {default as Navbar} from './navbar'
 export {default as Footer} from './footer'
 export {default as UserHome} from './user-home'
 export {Login, Signup} from './auth-form'
-export {Billing, Shipping} from './order-form'
+export {Billing, Shipping} from './cart/order-form'
